refactor: Board를 프린트하는 구조 및 로직 변경
