+++ README.md
@@ -61,3 +61,10 @@
 
 - [x] 초기 턴은 WHITE부터 시작된다.
 - [x] WHITE에서 BLACK으로, BLACK에서 WHITE로 턴을 차례로 바꾼다.
+
+
+### Score
+- [ ] 현재까지 남아 있는 말에 따라 점수를 계산할 수 있어야 한다.
+- [ ] 각 말의 점수는 queen은 9점, rook은 5점, bishop은 3점, knight는 2.5점이다.
+- [ ] pawn의 기본 점수는 1점이다. 
+  - [ ] 하지만 같은 세로줄에 같은 색의 폰이 있는 경우 1점이 아닌 0.5점을 준다.
