+++ src/main/java/chess/domain/game/Position.java
@@ -4,9 +4,6 @@ import java.util.Objects;
 
 public final class Position {
 
-    private static final int ONE_SPACE = 1;
-    private static final int TWO_SPACES = 2;
-
     private final File file;
     private final Rank rank;
